requirements.txt
git+https://github.com/fabfic/ether.git@master#egg=edgerun-ether
edgerun-skippy-core>=0.1.1
simpy==3.0.11
scipy==1.7.0
numpy==1.22.0
matplotlib==3.5.0
pandas==1.2.5
srds==0.1.0
scikit-learn==0.24.2
joblib==0.17.0
tpot==0.11.5
tqdm
requests
